feat!(blueprint-test-utils): moved eigen test to blueprint and improving eigen context wip
