+++ blueprint-test-utils/src/lib.rs
@@ -39,14 +39,12 @@ pub type InputValue = runtime_types::tangle_primitives::services::field::Field<A
 pub type OutputValue = runtime_types::tangle_primitives::services::field::Field<AccountId32>;
 
 pub mod anvil;
-pub mod helpers;
-pub mod sync;
-pub mod test_ext;
-
-#[cfg(feature = "eigenlayer_test")]
 pub mod eigenlayer_test_env;
+pub mod helpers;
 pub mod incredible_squaring_helpers;
 pub mod symbiotic_test_env;
+pub mod sync;
+pub mod test_ext;
 
 pub type TestClient = TangleClient;
