+++ sdk/src/ctx.rs
@@ -32,15 +32,20 @@
 //! ```
 //!
 
-use core::future::Future;
-
 use crate::keystore::backend::GenericKeyStore;
+use alloy_primitives::{Address, Bytes, FixedBytes, U256};
+use core::future::Future;
+use eigensdk::types::operator::OperatorPubKeys;
+use eigensdk::utils::binding::DelegationManager::OperatorDetails;
+use eigensdk::utils::binding::OperatorStateRetriever;
+use eigensdk::utils::binding::StakeRegistry::StakeUpdate;
 use eigensdk::{
     client_avsregistry::{reader::AvsRegistryChainReader, writer::AvsRegistryChainWriter},
     services_avsregistry::chaincaller::AvsRegistryServiceChainCaller,
     services_blsaggregation::bls_agg::BlsAggregatorService,
     services_operatorsinfo::operatorsinfo_inmemory::OperatorInfoServiceInMemory,
 };
+use std::collections::HashMap;
 // derives
 pub use gadget_context_derive::*;
 use tangle_subxt::tangle_testnet_runtime::api::runtime_types::{
@@ -137,4 +142,83 @@ pub trait EigenlayerContext {
         >,
         std::io::Error,
     >;
+
+    /// Get Operator stake in Quorums at a given block.
+    async fn get_operator_stake_in_quorums_at_block(
+        &self,
+        block_number: u32,
+        quorum_numbers: Bytes,
+    ) -> Result<Vec<Vec<OperatorStateRetriever::Operator>>, std::io::Error>;
+
+    /// Get an Operator's stake in Quorums at current block.
+    async fn get_operator_stake_in_quorums_at_current_block(
+        &self,
+        operator_id: FixedBytes<32>,
+    ) -> Result<HashMap<u8, U256>, std::io::Error>;
+
+    /// Get an Operator by ID.
+    async fn get_operator_by_id(&self) -> Result<OperatorStateRetriever::Operator, std::io::Error>;
+
+    /// Get an Operator stake history.
+    async fn get_operator_stake_history(
+        &self,
+        operator_id: FixedBytes<32>,
+        quorum_number: u8,
+    ) -> Result<Vec<StakeUpdate>, std::io::Error>;
+
+    /// Get an Operator stake update at a given index.
+    async fn get_operator_stake_update_at_index(
+        &self,
+        operator_id: FixedBytes<32>,
+        quorum_number: u8,
+        index: U256,
+    ) -> Result<StakeUpdate, std::io::Error>;
+
+    /// Get an Operator's stake at a given block number.
+    async fn get_operator_stake_at_block_number(
+        &self,
+        operator_id: FixedBytes<32>,
+        quorum_number: u8,
+        block_number: u32,
+    ) -> Result<U256, std::io::Error>;
+
+    /// Get an Operator's [`details`](OperatorDetails).
+    async fn get_operator_details(
+        &self,
+        operator_addr: Address,
+    ) -> Result<OperatorDetails, std::io::Error>;
+
+    /// Get an Operator's latest stake update.
+    async fn get_latest_stake_update(
+        &self,
+        operator_id: FixedBytes<32>,
+        quorum_number: u8,
+    ) -> Result<StakeUpdate, std::io::Error>;
+
+    /// Get an Operator's ID as [`FixedBytes`] from its [`Address`].
+    async fn get_operator_id(
+        &self,
+        operator_addr: Address,
+    ) -> Result<FixedBytes<32>, std::io::Error>;
+
+    /// Get the total stake at a given block number from a given index.
+    async fn get_total_stake_at_block_number_from_index(
+        &self,
+        quorum_number: u8,
+        block_number: u32,
+        index: U256,
+    ) -> Result<U256, std::io::Error>;
+
+    /// Get the total stake history length of a given quorum.
+    async fn get_total_stake_history_length(
+        &self,
+        quorum_number: u8,
+    ) -> Result<U256, std::io::Error>;
+
+    /// Provides the public keys of existing registered operators within the provided block range.
+    async fn query_existing_registered_operator_pub_keys(
+        &self,
+        start_block: u32,
+        to_block: u32,
+    ) -> Result<Vec<OperatorPubKeys>, std::io::Error>;
 }
